+++ src/imap_stream.rs
@@ -1,9 +1,9 @@
 use std::fmt;
 use std::pin::Pin;
 
-#[cfg(feature = "runtime-async-std")]
-use async_std::io::{Read, Write, WriteExt};
 use bytes::BytesMut;
+#[cfg(not(feature = "runtime-tokio"))]
+use futures_util::io::{AsyncRead as Read, AsyncWrite as Write, AsyncWriteExt};
 use futures_util::stream::Stream;
 use futures_util::task::{Context, Poll};
 use futures_util::{io, ready};
@@ -162,7 +162,7 @@ impl<R: Read + Write + Unpin> ImapStream<R> {
             // even if it is called with 0 as an argument.
             debug_assert!(!buf.is_empty());
 
-            #[cfg(feature = "runtime-async-std")]
+            #[cfg(not(feature = "runtime-tokio"))]
             let num_bytes_read = ready!(Pin::new(&mut this.inner).poll_read(cx, buf))?;
 
             #[cfg(feature = "runtime-tokio")]
@@ -389,13 +389,13 @@ mod tests {
         }
     }
 
-    #[cfg(feature = "runtime-async-std")]
+    #[cfg(not(feature = "runtime-tokio"))]
     impl Read for FailingStream {
         fn poll_read(
             self: Pin<&mut Self>,
             cx: &mut Context<'_>,
             buf: &mut [u8],
-        ) -> Poll<async_std::io::Result<usize>> {
+        ) -> Poll<std::io::Result<usize>> {
             let this = self.project();
             if !*this.has_failed {
                 *this.has_failed = true;
@@ -429,27 +429,21 @@ mod tests {
         }
     }
 
-    #[cfg(feature = "runtime-async-std")]
+    #[cfg(not(feature = "runtime-tokio"))]
     impl Write for FailingStream {
         fn poll_write(
             self: Pin<&mut Self>,
             _cx: &mut Context<'_>,
             buf: &[u8],
-        ) -> Poll<async_std::io::Result<usize>> {
+        ) -> Poll<std::io::Result<usize>> {
             Poll::Ready(Ok(buf.len()))
         }
 
-        fn poll_flush(
-            self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<async_std::io::Result<()>> {
+        fn poll_flush(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
             Poll::Ready(Ok(()))
         }
 
-        fn poll_close(
-            self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<async_std::io::Result<()>> {
+        fn poll_close(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
             Poll::Ready(Ok(()))
         }
     }
@@ -463,6 +457,7 @@ mod tests {
     /// or returning an inifinite stream of errors.
     #[cfg_attr(feature = "runtime-tokio", tokio::test)]
     #[cfg_attr(feature = "runtime-async-std", async_std::test)]
+    #[cfg_attr(feature = "runtime-futures", async_std::test)]
     async fn test_imap_stream_error() {
         use futures_util::StreamExt;