+++ src/lib.rs
@@ -76,11 +76,23 @@
 #![warn(missing_docs)]
 #![deny(rust_2018_idioms, unsafe_code)]
 
-#[cfg(not(any(feature = "runtime-tokio", feature = "runtime-async-std")))]
-compile_error!("one of 'runtime-async-std' or 'runtime-tokio' features must be enabled");
+#[cfg(not(any(
+    feature = "runtime-tokio",
+    feature = "runtime-async-std",
+    feature = "runtime-futures"
+)))]
+compile_error!(
+    "one of 'runtime-async-std', 'runtime-tokio' or 'runtime-futures' features must be enabled"
+);
 
-#[cfg(all(feature = "runtime-tokio", feature = "runtime-async-std"))]
-compile_error!("only one of 'runtime-async-std' or 'runtime-tokio' features must be enabled");
+#[cfg(any(
+    all(feature = "runtime-tokio", feature = "runtime-async-std"),
+    all(feature = "runtime-tokio", feature = "runtime-futures"),
+    all(feature = "runtime-async-std", feature = "runtime-futures"),
+))]
+compile_error!(
+    "only one of 'runtime-async-std', 'runtime-tokio' or 'runtime-futures' features must be enabled"
+);
 #[macro_use]
 extern crate pin_utils;